Add possible structure for moludes
