+++ js/index.js
@@ -1,107 +1,108 @@
+import { Game } from "./modules/game.js";
+
 window.onload = () => {
   const board = document.getElementById('board');
   board.hidden = true;
 };
 
-const gameBoard = ['', '', '', '', '', '', '', '', ''];
+// const gameBoard = ['', '', '', '', '', '', '', '', ''];
 
-const Game = (players) => {
-  let currentPlayer = players[0];
-  const finished = false;
-  let number_cells_selected = 0;
+// const Game = (players) => {
+//   let currentPlayer = players[0];
+//   let numberCellsSelected = 0;
 
-  const start = () => {
-    assignCells();
-    displayPlayerTurn(currentPlayer);
-  };
+//   const hidePlayerTurn = () => {
+//     const playerTurnLabel = document.getElementById('player-turn-label');
+//     playerTurnLabel.hidden = true;
+//   };
 
-  const checkWinner = () => {
-    const p1_moves = [];
-    const p2_moves = [];
+//   const disableButtons = () => {
+//     const cells = document.getElementsByClassName('cell');
+//     for (let i = 0; i < cells.length; i += 1) {
+//       cells[i].disabled = true;
+//     }
+//   };
 
-    const winning_combinations = [[0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 3, 6],
-      [1, 4, 7],
-      [2, 5, 8],
-      [0, 4, 8],
-      [2, 4, 6]];
+//   const checkWinner = () => {
+//     const p1Moves = [];
+//     const p2Moves = [];
 
-    gameBoard.forEach((val, idx) => {
-      if (val == 'X') p1_moves.push(idx);
-      if (val == 'O') p2_moves.push(idx);
-    });
+//     const winningCombinations = [[0, 1, 2],
+//       [3, 4, 5],
+//       [6, 7, 8],
+//       [0, 3, 6],
+//       [1, 4, 7],
+//       [2, 5, 8],
+//       [0, 4, 8],
+//       [2, 4, 6]];
 
-    for (let i = 0; i < winning_combinations.length; i++) {
-      const reg = new RegExp(winning_combinations[i].join('.*'));
-      const winnerResult = document.getElementById('game-result');
+//     gameBoard.forEach((val, idx) => {
+//       if (val === 'X') p1Moves.push(idx);
+//       if (val === 'O') p2Moves.push(idx);
+//     });
 
-      if (reg.test(p1_moves.join(''))) {
-        winnerResult.innerHTML = `${players[0].name} has won`;
-        disableButtons();
-        hidePlayerTurn();
-        break;
-      } else if (reg.test(p2_moves.join(''))) {
-        winnerResult.innerHTML = `${players[1].name} has won`;
-        disableButtons();
-        hidePlayerTurn();
-        break;
-      } else if (number_cells_selected == 9) {
-        winnerResult.innerHTML = "It's a draw";
-        hidePlayerTurn();
-      }
-    }
-  };
+//     for (let i = 0; i < winningCombinations.length; i += 1) {
+//       const reg = new RegExp(winningCombinations[i].join('.*'));
+//       const winnerResult = document.getElementById('game-result');
 
-  const selectCell = () => {
-    const { id } = event.target;
-    el = document.getElementById(id);
-    el.innerHTML = currentPlayer.symbol;
-    gameBoard[id[3]] = el.innerHTML;
-    number_cells_selected++;
-    el.disabled = true;
-    checkWinner();
-    changePlayerTurn();
-  };
+//       if (reg.test(p1Moves.join(''))) {
+//         winnerResult.innerHTML = `${players[0].name} has won`;
+//         disableButtons();
+//         hidePlayerTurn();
+//         break;
+//       } else if (reg.test(p2Moves.join(''))) {
+//         winnerResult.innerHTML = `${players[1].name} has won`;
+//         disableButtons();
+//         hidePlayerTurn();
+//         break;
+//       } else if (numberCellsSelected === 9) {
+//         winnerResult.innerHTML = "It's a draw";
+//         hidePlayerTurn();
+//       }
+//     }
+//   };
 
-  const changePlayerTurn = () => {
-    currentPlayer = currentPlayer.id == 1 ? players[1] : players[0];
-    displayPlayerTurn(currentPlayer);
-  };
+//   const displayPlayerTurn = (currentPlayer) => {
+//     const playerTurnLabel = document.getElementById('player-turn-label');
+//     playerTurnLabel.innerHTML = `${currentPlayer.name}'s turn`;
+//   };
 
-  const displayPlayerTurn = (currentPlayer) => {
-    const playerTurnLabel = document.getElementById('player-turn-label');
-    playerTurnLabel.innerHTML = `${currentPlayer.name}'s turn`;
-  };
+//   const changePlayerTurn = () => {
+//     currentPlayer = currentPlayer.id === 1 ? players[1] : players[0];
+//     displayPlayerTurn(currentPlayer);
+//   };
 
-  const hidePlayerTurn = () => {
-    const playerTurnLabel = document.getElementById('player-turn-label');
-    playerTurnLabel.hidden = true;
-  };
+//   const selectCell = () => {
+//     const { id } = event.target; // eslint-disable-line no-restricted-globals
+//     const el = document.getElementById(id);
+//     el.innerHTML = currentPlayer.symbol;
+//     gameBoard[id[3]] = el.innerHTML;
+//     numberCellsSelected += 1;
+//     el.disabled = true;
+//     checkWinner();
+//     changePlayerTurn();
+//   };
 
-  const assignCells = () => {
-    for (let i = 0; i < gameBoard.length; i++) {
-      const cell = document.getElementById(`pos${i}`);
-      cell.onclick = selectCell.bind(cell);
-    }
-  };
+//   const assignCells = () => {
+//     for (let i = 0; i < gameBoard.length; i += 1) {
+//       const cell = document.getElementById(`pos${i}`);
+//       cell.onclick = selectCell.bind(cell);
+//     }
+//   };
 
-  const disableButtons = () => {
-    const cells = document.getElementsByClassName('cell');
-    for (let i = 0; i < cells.length; i++) {
-      cells[i].disabled = true;
-    }
-  };
+//   const start = () => {
+//     assignCells();
+//     displayPlayerTurn(currentPlayer);
+//   };
 
-  return {
-    start, players, selectCell, currentPlayer, changePlayerTurn,
-  };
-};
+//   return {
+//     start, players, selectCell, currentPlayer, changePlayerTurn,
+//   };
+// };
 
 const Player = (id, name, symbol) => ({ id, name, symbol });
 
-const initGame = () => {
+const initGame = () => { // eslint-disable-line no-unused-vars
   const player1Name = document.getElementById('player-one').value;
   const player2Name = document.getElementById('player-two').value;
 
@@ -117,9 +118,11 @@ const initGame = () => {
   board.hidden = false;
   const cells = document.getElementsByClassName('cell');
 
-  for (let i = 0; i < cells.length; i++) {
+  for (let i = 0; i < cells.length; i += 1) {
     cells[i].disabled = false;
   }
 
   game.start();
 };
+
+export { initGame };
